refactor(selection-actions): extract CSS and stream logic from content script
- overlay-shadow-styles.ts: all shadow DOM CSS (was inline template literal)
- overlay-stream.ts: chrome port connection and chunk/done/error handling
- content script drops from 966 to 534 lines, pure orchestration
